pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-cloud"
version = "0.3.0"
description = "Hardware-accelerated tokenomics wind tunnel for DePIN protocols"
readme = "README.md"
authors = [
    {name = "Borja Tarazona Raga", email = "borja@sentinelcloud.dev"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Testing",
]
keywords = ["depin", "tokenomics", "simulation", "fpga", "hardware", "settlement"]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "click>=8.0.0",
    "pydantic>=1.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
dashboard = [
    "streamlit>=1.20.0",
    "plotly>=5.0.0",
]
hardware = [
    "cocotb==1.8.1",
]

[project.urls]
Homepage = "https://github.com/BorjaTR/Sentinel_Lab"
Documentation = "https://github.com/BorjaTR/Sentinel_Lab#readme"
Repository = "https://github.com/BorjaTR/Sentinel_Lab"
"Bug Tracker" = "https://github.com/BorjaTR/Sentinel_Lab/issues"

[project.scripts]
sentinel = "sentinel_cloud.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["sentinel_cloud*"]
exclude = ["tests*", "tb*", "rtl*", "model*"]

[tool.setuptools.package-data]
sentinel_cloud = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | sim_build
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=sentinel_cloud --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
